+++ Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [UNRELEASED]
 ### Fixed
  - Conditionals for Form component.
+ - Token issue for iframe.
 
 ## 3.5.7
 ### Fixed